Prevent float64 in threshold and weighted blending

+++ lib/improver/blending/weights.py
@@ -300,7 +300,7 @@ class ChooseDefaultWeightsLinear(object):
         """
         # Special case num_of_weighs == 1 i.e. Scalar coordinate.
         if num_of_weights == 1:
-            weights = np.array([1.0])
+            weights = np.array([1.0], dtype=np.float32)
             return weights
         if not isinstance(self.y0val, float) or self.y0val < 0.0:
             msg = ('y0val must be a float >= 0.0, '
@@ -318,7 +318,8 @@ class ChooseDefaultWeightsLinear(object):
         for tval in range(0, num_of_weights):
             weights_list.append(self.slope*tval + self.y0val)
 
-        weights = WeightsUtilities.normalise_weights(np.array(weights_list))
+        weights = WeightsUtilities.normalise_weights(
+            np.array(weights_list, dtype=np.float32))
 
         return weights
 
@@ -412,7 +413,8 @@ class ChooseDefaultWeightsNonLinear(object):
         for tval_minus1 in range(0, num_of_weights):
             weights_list.append(self.cval**(tval_minus1))
 
-        weights = WeightsUtilities.normalise_weights(np.array(weights_list))
+        weights = WeightsUtilities.normalise_weights(
+            np.array(weights_list, dtype=np.float32))
 
         return weights
 
@@ -523,7 +525,7 @@ class ChooseDefaultWeightsTriangular(object):
             return weight
 
         slope = 1.0/width
-        weights = np.zeros(coord_vals.shape)
+        weights = np.zeros(coord_vals.shape, dtype=np.float32)
         # Find the indices of the points where there will be non-zero weights.
         condition = ((coord_vals >= (midpoint-width)) &
                      (coord_vals <= (midpoint+width)))
